+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
  - Added a method for converting 24 hour to 12 hour format
+ - Added a method for determining if a culture uses 24 hour format
 
 ## [9.8.0] - 2019-09-03
 ### Added